+++ docs/web_modules/Footer/index.js
@@ -1,4 +1,5 @@
 import React, { Component } from "react"
+import { Link } from "react-router"
 
 import styles from "./index.css"
 
@@ -8,7 +9,7 @@ export default class Footer extends Component {
     return (
       <footer className={ styles.footer }>
         <a
-          href="http://moox.io/statinamic/"
+          href="http://moox.io/"
           className={ styles.link }
         >
           { "Cooked by " }
@@ -20,6 +21,10 @@ export default class Footer extends Component {
         <a href="https://github.com/MoOx/statinamic">
           { "Source on GitHub" }
         </a>
+        { " | " }
+        <Link to="/statinamic/404.html">
+          { "404" }
+        </Link>
       </footer>
     )
   }