+++ docs/web_modules/layouts/PageError/index.js
@@ -4,8 +4,13 @@ import { PropTypes } from "react"
 export default class PageError extends Component {
 
   static propTypes = {
-    error: PropTypes.number.isRequired,
-    errorText: PropTypes.string.isRequired,
+    error: PropTypes.number,
+    errorText: PropTypes.string,
+  };
+
+  static defaultProps = {
+    error: 404,
+    errorText: "Page Not Found",
   };
 
   render() {