+++ src/static/to-html/__tests__/url-as-html.js
@@ -2,8 +2,8 @@ import test from "ava"; import "babel-core/register"
 
 import url from "url"
 
-import beautifyHTML from "../../../__tests__/utils/beautifyHTML"
-import htmlMetas from "../../../html-metas"
+import beautifyHTML from "../../../_utils/beautify-html"
+import htmlMetas from "../../../_utils/html-metas"
 
 import urlAsHtml from "../url-as-html"
 
@@ -22,14 +22,14 @@ const fixture = {
 }
 
 test("url as html", async (t) => {
-  urlAsHtml("", fixture, true)
+  urlAsHtml("/", fixture, true)
   .then((html) => {
     const expectedHTML = (
 `<!doctype html>
 <html lang="en">
 
 <head>
-  ${ htmlMetas({ baseUrl: { path: "/" } }).join("\n  ") }
+  ${ htmlMetas({ baseUrl: { pathname: "/" } }).join("\n  ") }
   <title data-react-helmet="true"></title>
 </head>